requirements.txt
attrs==15.2.0
Automat==0.0.0
cffi==1.9.1
click==6.7
constantly==15.1.0
cryptography==1.7.1
cssselect==1.0.0
Cython==0.25.2
docker==2.2.1
docker-py==1.10.6
docker-pycreds==0.2.1
idna==2.2
incremental==16.10.1
lxml==3.7.3
olefile==0.44
paramiko==2.1.2
parsel==1.1.0
Pillow==4.1.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pytesseract==0.1.6
pytz==2017.2
pywin32==220
PyYAML==3.12
queuelib==1.4.2
retrying==1.3.3
scrapinghub==2.0.0
Scrapy==1.3.3
scrapy-splash==0.7.2
service-identity==16.0.0
shub==2.5.1
six==1.10.0
tesseract==0.1.3
Twisted==17.1.0
w3lib==1.17.0
websocket-client==0.40.0
zope.interface==4.3.3
